+++ CHANGELOG.rst
@@ -2,6 +2,7 @@
 =====
 
 * Added support for returning multiple objects. (`#7`_)
+* Refactored models instantiation. (`#9`_)
 
 0.1.2
 =====
@@ -19,5 +20,6 @@
 * Initial release.
 
 
+.. _#9: https://github.com/Stranger6667/pyoffers/issues/9
 .. _#7: https://github.com/Stranger6667/pyoffers/issues/7
 .. _#5: https://github.com/Stranger6667/pyoffers/issues/5